+++ Backend/src/main/java/com/example/backend/service/impl/DanhMucServiceImpl.java
@@ -1,7 +1,9 @@
 package com.example.backend.service.impl;
 
 import com.example.backend.entity.DanhMuc;
+import com.example.backend.exception.BadRequestException;
 import com.example.backend.repository.DanhMucRepository;
+import com.example.backend.repository.SanPhamRepository;
 import com.example.backend.service.DanhMucService;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -14,6 +16,9 @@ public class DanhMucServiceImpl implements DanhMucService {
 
     @Autowired
     private DanhMucRepository danhMucRepository;
+
+    @Autowired
+    private SanPhamRepository sanPhamRepository;
 
     @Override
     public List<DanhMuc> getAll() {
@@ -32,6 +37,9 @@ public class DanhMucServiceImpl implements DanhMucService {
 
     @Override
     public void delete(Integer id) {
+        if (sanPhamRepository.existsByMaDanhMuc(id)) {
+            throw new BadRequestException("Không thể xóa danh mục này vì đang có sản phẩm thuộc danh mục!");
+        }
         danhMucRepository.deleteById(id);
     }
 }
